refactor(cursor): move cursor helpers from perf/ to utils/
